pyproject.toml
[project]
name = "amc-backend"
version = "0.1.0"
description = "ASEAN Motor Club backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.3",
    "django-ninja>=1.4.3",
    "gunicorn>=23.0.0",
    "psycopg[binary]>=3.2.9",
    "uvicorn[standard]>=0.34.3",
    "uvicorn-worker>=0.3.0",
    "aiohttp>=3.12.13",
    "django-asgi-lifespan>=0.4.0",
    "arq>=0.26.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.taskipy.tasks]
dev = "uvicorn amc_backend.asgi:application --reload --lifespan=on"
test = "django-admin test"

[tool.basedpyright]
allowedUntypedLibraries = [
  "django"
]

[tool.hatch.build.targets.wheel]
sources = ["src"]
packages = [
  "src/amc_backend",
  "src/amc",
]

[dependency-groups]
dev = [
    "basedpyright>=1.29.5",
    "taskipy>=1.14.1",
]

